feat: QRDecoder 리팩토링 반영 및 QR 정보 React 뷰 개선
