+++ src/features/decode/types/detailProps.ts
@@ -1,9 +1,4 @@
-import type { QRDecodeResult } from "@/libs/QRDecoder/types/QRDecodeResult";
-
-import { QRDecoder } from "@/libs/QRDecoder";
-
 export interface DetailProps {
   color?: string | undefined;
-  qrDecoder: QRDecoder;
-  qrDecodeResult: QRDecodeResult;
+  matrix: number[][];
 }
